Move `CoreDataFeedStore+FeedStore` to it's own file
