+++ FeedFetcher/FeedFetcher/FeedCache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -15,63 +15,13 @@ public final class CoreDataFeedStore {
         context = container.newBackgroundContext()
     }
             
-    private func perform(_ action:@escaping (NSManagedObjectContext) -> Void) {
+    func perform(_ action:@escaping (NSManagedObjectContext) -> Void) {
         let context = self.context
         context.perform {
             action(context)
         }
     }
                 
-}
-
-// MARK:- FeedStore
-
-extension CoreDataFeedStore: FeedStore {
-    
-    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        perform { context in
-            completion(Result {
-                try CoreDataFeedStore.deleteAllManagedCache(in: context)
-                try context.saveIfHasPendingChanges()
-            })
-        }
-    }
-    
-    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        perform { context in
-            completion(Result {
-                try CoreDataFeedStore.deleteAllManagedCache(in: context)
-                
-                ManagedCache.mapFrom((timestamp: timestamp, images: feed), in: context)
-                try context.saveIfHasPendingChanges()
-            })
-        }
-    }
-    
-    public func retrieveCachedFeed(completion: @escaping RetrievalCompletion) {
-        perform { context in
-            completion(Result {
-                guard let managedCache = try Self.fetchFirstManagedCache(from: context) else {
-                    return nil
-                }
-                
-                let retrieveResult = managedCache.mapTo()
-                return CachedFeed(feed: retrieveResult.feed, timestamp: retrieveResult.timestamp)
-            })
-        }
-    }
-    
-    private static func deleteAllManagedCache(in context: NSManagedObjectContext) throws {
-        try fetchFirstManagedCache(from: context).flatMap { context.delete($0) }
-    }
-    
-    @discardableResult
-    private static func fetchFirstManagedCache(from context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
-        
-        return try context.fetch(request).first
-    }
-    
 }
 
 // MARK:- FeedImageDataStore
